编译问题修复

+++ src/platform/hal/hal_sysfreq.c
@@ -323,7 +323,7 @@ WEAK SYSTEM_SYSFREQ_STAT_T* system_sysfreq_stat_get(void)
     return &system_sysfreq_stat_info;
 }
 
-WEAK void system_sysfreq_stat_update(uint32_t tot, uint32_t arrFreq[][2], uint8_t size)
+WEAK void system_sysfreq_stat_update(uint32_t tot, uint32_t arrFreq[][2], uint32_t size)
 {
     system_sysfreq_stat_info.total_intvl = tot;
     memcpy((uint8_t*)(system_sysfreq_stat_info.sysfreq_intvl), (uint8_t*)arrFreq, size);
